Added KTSlidingWindowFFT, which uses a window function to produce windowed FFTs. The window can move along the event, producing a power spectrum at each point. The FFT class can create a waterfall plot of the result.
Added PlotPowerSpectrum.C to plot a power spectrum, along with the
amplitude and power distributions.

Many fixes, in particular to the window function classes.

Standardized some notation: # bins is called "Size"; "Length" is the
Size*BinWidth.  This is now consistently used for the time-domain data,
frequency-domain data, and window functions.

Caveat: in the plotting of the power distribution, the peaks are not cut
out, so the distribution includes the very-high-value peak bins.