KTEggProcessor handling of normalized time series
KTEggProcessor has been modified to handle the case in which
HatchNextSlice returns data either in TSRawTimeSeries (raw data) or
TSTimeSeries (normalized data).  If it is already normalized, then it
skips the NormalizeData function.  This is necessary because the RSA
stores the data already in voltage (that is, normalized), so that
RSAMATReader creates only TSTimeSeries and not TSRawTimeSeries.