Added TimeInAcq to KTSliceHeader
Added a field called TimeInAcq to KTSliceHeader, to store the slice
time relative to the current acquisition, instead of the beginning of
the run.  This distinction is necessary for triggered data, in which a
run can contain many acquisitions, and we need the TimeInAcq to place
an event relative to the trigger.  Also added calls to populate it in
KTRSAMatReader and KTEggReaderMonarch and 2011 (it’s identical to
TimeInRun in the latter two)