Changed HDF5 FFT writer to save all data into single dataset
Modified the HDF5 FFT writer (KTHDF5TypeWriterTransform) to save all
the data from an Egg or MAT file into a single HDF5 dataset, instead of
saving one dataset per slice.  We also streamlined the writer processor
- now it creates a single buffer and dataset with the required
dimensions, instead of creating several buffers and datasets for each
possible output, and then choosing which one to use.A future iteration
could be made to save different records as separate datasets.  We also
fixed references to Channel Headers, which were incorrect and causing
the HDF5 writer to crash. Finally, we noticed that the number of
channels is being set to 2 when it should be 1.  We added some debug
messages to try to track down the problem, but could not find where the
number of channels is being incremented.